bugfix: set social sharing image to article main image on article pages
